refactor btns into folder
