+++ src/components/buttons/DeleteEventBtn.tsx
