+++ src/components/modals/DeleteEventModal.tsx
